refactored auth to security
