+++ shakedown/__init__.py
@@ -1,7 +1,6 @@
 from shakedown.cli import *
 from shakedown.cli.helpers import *
 from shakedown.dcos import *
-from shakedown.dcos.auth import *
 from shakedown.dcos.cluster import *
 from shakedown.dcos.command import *
 from shakedown.dcos.config import *
@@ -9,6 +8,7 @@ from shakedown.dcos.file import *
 from shakedown.dcos.marathon import *
 from shakedown.dcos.network import *
 from shakedown.dcos.package import *
+from shakedown.dcos.security import *
 from shakedown.dcos.service import *
 from shakedown.dcos.spinner import *
 from shakedown.dcos.task import *
